chore: rename lib/gui/styles.h => lib/gui/Styles.h
